removed the test setup file and service worker file and moved components to its own folder
