--- src/CardSources.js
@@ -1,118 +1,0 @@
-//TODO: Find a better way of doing this
-import c1 from "./cardImages/c1.svg"; //1
-import c2 from "./cardImages/c2.svg"; //2
-import c3 from "./cardImages/c3.svg"; //3
-import c4 from "./cardImages/c4.svg"; //4
-import c5 from "./cardImages/c5.svg"; //5
-import c6 from "./cardImages/c6.svg"; //6
-import c7 from "./cardImages/c7.svg"; //7
-import c8 from "./cardImages/c8.svg"; //8
-import c9 from "./cardImages/c9.svg"; //9
-import c10 from "./cardImages/c10.svg"; // 10
-import c11 from "./cardImages/c11.svg"; // 11
-import c12 from "./cardImages/c12.svg"; // 12
-import c13 from "./cardImages/c13.svg"; // 13
-import h1 from "./cardImages/h1.svg"; // 14
-import h2 from "./cardImages/h2.svg"; // 15
-import h3 from "./cardImages/h3.svg"; // 16
-import h4 from "./cardImages/h4.svg"; // 17
-import h5 from "./cardImages/h5.svg"; // 18
-import h6 from "./cardImages/h6.svg"; // 19
-import h7 from "./cardImages/h7.svg"; // 20
-import h8 from "./cardImages/h8.svg"; // 21
-import h9 from "./cardImages/h9.svg"; // 22
-import h10 from "./cardImages/h10.svg"; // 23
-import h11 from "./cardImages/h11.svg"; // 24
-import h12 from "./cardImages/h12.svg"; // 25
-import h13 from "./cardImages/h13.svg"; // 26
-import s1 from "./cardImages/s1.svg"; // 27
-import s2 from "./cardImages/s2.svg"; // 28
-import s3 from "./cardImages/s3.svg"; // 29
-import s4 from "./cardImages/s4.svg"; // 30
-import s5 from "./cardImages/s5.svg"; // 31
-import s6 from "./cardImages/s6.svg"; // 32
-import s7 from "./cardImages/s7.svg"; // 33
-import s8 from "./cardImages/s8.svg"; // 34
-import s9 from "./cardImages/s9.svg"; // 35
-import s10 from "./cardImages/s10.svg"; // 36
-import s11 from "./cardImages/s11.svg"; // 37
-import s12 from "./cardImages/s12.svg"; // 38
-import s13 from "./cardImages/s13.svg"; // 39
-import d1 from "./cardImages/d1.svg"; // 40
-import d2 from "./cardImages/d2.svg"; // 41
-import d3 from "./cardImages/d3.svg"; // 42
-import d4 from "./cardImages/d4.svg"; // 43
-import d5 from "./cardImages/d5.svg"; // 44
-import d6 from "./cardImages/d6.svg"; // 45
-import d7 from "./cardImages/d7.svg"; // 46
-import d8 from "./cardImages/d8.svg"; // 47
-import d9 from "./cardImages/d9.svg"; // 48
-import d10 from "./cardImages/d10.svg"; // 49
-import d11 from "./cardImages/d11.svg"; // 50
-import d12 from "./cardImages/d12.svg"; // 51
-import d13 from "./cardImages/d13.svg"; // 52
-import j1 from "./cardImages/j1.svg"; // 53
-import j2 from "./cardImages/j2.svg"; // 54
-import b6 from "./cardImages/b6.svg"; // 54
-
-let CardSources = {
-  fronts: [
-    c1,
-    c2,
-    c3,
-    c4,
-    c5,
-    c6,
-    c7,
-    c8,
-    c9,
-    c10,
-    c11,
-    c12,
-    c13,
-    h1,
-    h2,
-    h3,
-    h4,
-    h5,
-    h6,
-    h7,
-    h8,
-    h9,
-    h10,
-    h11,
-    h12,
-    h13,
-    s1,
-    s2,
-    s3,
-    s4,
-    s5,
-    s6,
-    s7,
-    s8,
-    s9,
-    s10,
-    s11,
-    s12,
-    s13,
-    d1,
-    d2,
-    d3,
-    d4,
-    d5,
-    d6,
-    d7,
-    d8,
-    d9,
-    d10,
-    d11,
-    d12,
-    d13,
-    j1,
-    j2,
-  ],
-  backs: [b6],
-};
-
-export default CardSources;
